refactor: merge small landing sub-components into Widgets.jsx
